% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcdist.R
\name{arcdist}
\alias{arcdist}
\title{Distance between two coordinates}
\usage{
arcdist(lat1, lon1, lat2, lon2, R = 3959)
}
\arguments{
\item{lat1}{Latitude of first coordinate}

\item{lon1}{Longitude of first coordinate}

\item{lat2}{Latitude of second coordinate}

\item{lon2}{Longitude of second coordinate}

\item{R}{The radius of the Earth. Default is in miles, 3959. Change it to 6371 if you want the result in km.}
}
\description{
Put in point A and point B, get back miles or km.
Further info: http://janmatuschek.de/LatitudeLongitudeBoundingCoordinates
}
\keyword{distance}
\keyword{geocode}
\keyword{latitude}
\keyword{latlon}
\keyword{longitude}
