% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkify.R
\name{chunkify}
\alias{chunkify}
\title{Chunkify: Split Lists for Texting Upload}
\usage{
chunkify(data, chunk = 80000, filename)
}
\arguments{
\item{data}{Data frame or table}

\item{chunk}{Final list output size}

\item{filename}{Filenames for output. Don't need to include today's date or `part#` or `.csv`}
}
\description{
Chunkify: Split Lists for Texting Upload
}
