% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling.R
\name{clntxt}
\alias{clntxt}
\title{Clean text for easier grep}
\usage{
clntxt(text)
}
\arguments{
\item{text}{String to clean.}
}
\description{
Convert to lower, remove punctuation and special characters, trim whitespace, etc.
}
\details{
TODO: figure out how to remove emoji/unicode
}
\keyword{file}
\keyword{grep}
\keyword{grepl}
\keyword{parse}
