% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_utf.R
\docType{data}
\name{convert_utf}
\alias{convert_utf}
\alias{pycode}
\title{Convert UTF-16LE files to UTF-8}
\format{
An object of class \code{character} of length 15.
}
\usage{
convert_utf(infile, outfile, python_cmd="py")
}
\arguments{
\item{infile}{The path to the file to load (already extracted from zip).}

\item{outfile}{The name of the converted file to write (cannot be the same as infile).}

\item{python_cmd}{The path you use to get to python on your system's command line---common values are \code{py}, \code{python}, or \code{python3}.}
}
\description{
VAN changed its default export format. Boo!
}
\keyword{convert}
\keyword{utf}
\keyword{utf-16}
\keyword{utf-16le}
