% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.escape.R
\name{data.escape}
\alias{data.escape}
\title{Escape Data}
\usage{
data.escape(x)
}
\arguments{
\item{x}{Data to escape.}
}
\description{
Escape data for use in SQL queries.
Returns:
* NULL for NA,
* single quoted string for strings, factors, & dates,
* just the number for numbers.
}
\keyword{data}
\keyword{escape}
\keyword{query}
\keyword{sql}
