% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsafenames.R
\name{dbsafenames}
\alias{dbsafenames}
\title{Database-Safe Names}
\usage{
dbsafenames(x)
}
\arguments{
\item{x}{Whole data table or data frame.}
}
\description{
Make names db safe: no '.' or other illegal characters, all lower case and unique.
From: http://www.win-vector.com/blog/2016/02/using-postgresql-in-r/
}
