% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling.R
\name{fr}
\alias{fr}
\alias{rm_na_cols}
\title{Fast File Read 2}
\usage{
fr(filename, dbsafe=TRUE, progress = FALSE, ...)
}
\arguments{
\item{filename}{The name/location of the file, as a string. OR, a \code{list()} or \code{c()} of filenames, which will be bound together row-wise with an added column \code{file} that gives the filename it came from. NOTE: does use \code{glue()} on filenames by default!}

\item{dbsafe}{Clean column names using \code{workingamerica::dbsafenames()}?}

\item{progress}{Only used when reading a list of files, and used automatically.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Common settings for \code{fread()}. \code{...} lets you pass additional arguments to \code{data.table::fread()}. Built in support for reading and combining a list of files.
}
\keyword{file}
\keyword{fread}
\keyword{import}
\keyword{rm_na_cols}
