% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcolnames.R
\name{getcolnames}
\alias{getcolnames}
\title{Get names of columns in a db table}
\usage{
getcolnames(table_schema, table_name, cnxn=con)
}
\arguments{
\item{table_schema}{Schema}

\item{table_name}{Table}

\item{cnxn}{Connection, default con}
}
\description{
Get names of columns in a db table
}
\keyword{column}
\keyword{names}
\keyword{postgres}
\keyword{query}
\keyword{table}
