% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{randomize}
\alias{randomize}
\alias{randomize_cond}
\title{Automatic Householding and Randomization}
\usage{
randomize(data, household = T, maxhhsize = 5, conditions = c("Control", "Treatment"), 
  weight = rep(1/length(conditions), length(conditions)), verbose=FALSE)
}
\arguments{
\item{weight}{Numeric vector of the same length as the conditions vector, with weights assigned to each condition. Weights must add to 1. Must be in same order as conditions! Default is equal weights.}

\item{conditions}{Character vector of condition names. Default is c("Control", "Treatment")}

\item{verbose}{Default is FALSE, so it displays all diagnostics.}

\item{data}{Data frame or table}

\item{household}{True or False}

\item{maxhhsize}{Max household size (if household=TRUE); default 5. Larger "households" are excluded from the final randomized file on the suspicion of being uncanvassable apartment buildings.}
}
\description{
Automatic Householding and Randomization
}
