% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timed_insert.R
\name{timed_insert}
\alias{timed_insert}
\title{Simplified dbWriteTable()}
\usage{
timed_insert(data, schema = NA, table, append = TRUE, temp = FALSE, cnxn = con)
}
\arguments{
\item{data}{The data, as something coercible to a data.frame. If it's a data.frame already, the variables should be named identically to the table you're inserting into.}

\item{schema}{The schema to insert to. For temporary tables, use \code{NA}.}

\item{table}{The table to insert to.}

\item{append}{Quoting \code{dbWriteTable()} help: "If the append argument is TRUE, the rows in an existing table are preserved, and the new data are appended. If the table doesn't exist yet, it is created." Default \code{TRUE}.}

\item{temp}{Create temporary table? If so, schema will be set to \code{NA} and temporary table will be named to parameter \code{table}'s value. Warning: calling a temp table the same name as an existing table can cause unexpected results!}

\item{cnxn}{Connection, defaults to a global variable named \code{con}.}
}
\description{
Wrapper for \code{DBI::dbWriteTable()} with some typical defaults.
}
