% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timed_insert.R
\name{timed_refresh}
\alias{timed_refresh}
\title{Simplified materialized view refresh.}
\usage{
timed_refresh(schema, table, cnxn = con )
}
\arguments{
\item{schema}{The schema of the materialized view to refresh}

\item{table}{The table of the materialized view to refresh}

\item{cnxn}{Connection, defaults to a global variable named \code{con}.}
}
\description{
Times materialized view refreshes, and also clears results automatically.
}
