% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varchar.R
\name{varchar}
\alias{varchar}
\title{Check data coverage and max string length}
\usage{
varchar(x)
}
\arguments{
\item{x}{Whole data table or data frame.}
}
\description{
Find the max length & type & coverage of each column in a data.frame. Handy for when you're creating a new table.
}
\keyword{coverage}
\keyword{database}
\keyword{varchar}
