% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.colparse.R
\name{yaml.colparse}
\alias{yaml.colparse}
\alias{yaml.getvar}
\title{Vectorized YAML parsing}
\usage{
yaml.colparse(data, varname = c())
}
\arguments{
\item{varname}{String of the variable name you want to extract. If no varname is given, or if it's a vector of variable names, the function will instead print a list of the available variables. If the varname doesn't exist in a given row, function will return NA.}

\item{data}{YAML string---can be a vector of strings.}
}
\description{
Extract a single variable at a time from a field in a table that holds YAML data.
}
\examples{
donors <- data.table(
  id = c(1:6),
  yaml_values = c(
    "---\n- :type: Group\n  :id: 48\n  :amount: '16.25'\n",
    "---\n- :type: User\n  :id: 257\n  :amount: '10.00'\n",
    "---\n- :type: User\n  :id: 257\n  :amount: '15.00'\n",
    "---\n- :type: Group\n  :id: 48\n  :amount: '25.00'\n",
    "---\n- :type: Group\n  :id: 48\n  :amount: '10.00'\n",
    "---\n- :type: User\n  :id: 257\n  :amount: '15.00'\n"
  )
)
yaml.colparse(donors$yaml_values)
donors[, amount1 := yaml.colparse(yaml_values, ":amount") ]
donors[, amount2 := yaml.colparse(yaml_values, "amount") ]
donors
}
\keyword{action}
\keyword{actionnetwork}
\keyword{network}
\keyword{yaml}
