% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipthree.R
\name{zipthree}
\alias{zipthree}
\title{Zip3 extraction}
\usage{
zipthree(zip)
}
\arguments{
\item{zip}{A vector of (ostensible) zip codes.}
}
\description{
Zip3 extraction
}
\examples{
df <- data.frame(name = letters[1:6], 
  zip = c("4024","04011","20010","359-0000","80209-4444","90210"))
sapply(df$zip, zipthree) -> df$zip3
}
\keyword{3}
\keyword{Also}
\keyword{Even}
\keyword{TODO:}
\keyword{US}
\keyword{Zip}
\keyword{all}
\keyword{and}
\keyword{are}
\keyword{code}
\keyword{codes}
\keyword{digits}
\keyword{do}
\keyword{extract}
\keyword{first}
\keyword{has}
\keyword{if}
\keyword{import}
\keyword{impute}
\keyword{it's}
\keyword{kinds}
\keyword{leading}
\keyword{mapping}
\keyword{match.}
\keyword{messy,}
\keyword{missing}
\keyword{need}
\keyword{of}
\keyword{or}
\keyword{punctuation.}
\keyword{sometimes}
\keyword{state.}
\keyword{table}
\keyword{the}
\keyword{to}
\keyword{weird}
\keyword{you}
\keyword{zeroes}
\keyword{zip}
\keyword{zip3}
\keyword{zipthree}
